// frontend/src/types/generated/api/CreateOptimizationTaskRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MetaOptimizationTaskHint } from "./MetaOptimizationTaskHint";

export type CreateOptimizationTaskRequest = { name: string, description: string | null, goal: string, execution_target_type: string, task_mode: string, test_set_ids: Array<string>, meta_optimization: MetaOptimizationTaskHint | null, };